pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]

[tool.poetry]
name = "my_project"
version = "0.1.0"
description = "A sample project using Azure Cognitive Services."
authors = ["Thaddeus Thomas <thaddeus.r.thomas@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
azure-cognitiveservices-speech = "^1.32.1"
toml = "^0.10.2"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[tool.azure]
subscription_key = "YOUR_AZURE_SUBSCRIPTION_KEY"
region = "YOUR_AZURE_REGION"

[tool.my_project_settings]
output_directory = "./audio_output"
base_filename = "sample_audio"
